add skay to makefile and fixed type of neq

+++ kinsol.f90
@@ -8,7 +8,7 @@ use mkinsol
 implicit none
 
 integer ier
-integer*8 neq, i
+integer neq, i
 double precision udata(*), uscale(*), fdata(*), fscale(*)
 double precision vv(*), ftem(*)
 
@@ -35,7 +35,7 @@ use molecules
 implicit none
 
 integer ier
-integer*8 neq, i
+integer neq, i
 double precision udata(*), uscale(*), fdata(*), fscale(*)
 double precision vtemp1(*), vtemp2(*)
 integer n
